+++ traffic_ops/testing/api/v3/topologies_test.go
@@ -21,10 +21,11 @@ package v3
 
 import (
 	"fmt"
-	"github.com/apache/trafficcontrol/lib/go-tc"
 	"reflect"
 	"strings"
 	"testing"
+
+	"github.com/apache/trafficcontrol/lib/go-tc"
 )
 
 type topologyTestCase struct {
@@ -33,7 +34,7 @@ type topologyTestCase struct {
 }
 
 func TestTopologies(t *testing.T) {
-	WithObjs(t, []TCObj{Types, CacheGroups, CDNs, Parameters, Profiles, Statuses, Divisions, Regions, PhysLocations, Servers, Topologies}, func() {
+	WithObjs(t, []TCObj{Types, CacheGroups, CDNs, Parameters, Profiles, Statuses, Divisions, Regions, PhysLocations, Servers, ServerCapabilities, ServerServerCapabilitiesForTopologies, Topologies, Tenants, DeliveryServices, TopologyBasedDeliveryServiceRequiredCapabilities}, func() {
 		UpdateTestTopologies(t)
 		ValidationTestTopologies(t)
 		EdgeParentOfEdgeSucceedsWithWarning(t)
@@ -163,11 +164,10 @@ func updateSingleTopology(topology tc.Topology) error {
 }
 
 func UpdateTestTopologies(t *testing.T) {
-	topologiesCount := len(testData.Topologies)
-	for index := range testData.Topologies {
-		topology := testData.Topologies[(index+1)%topologiesCount]
-		topology.Name = testData.Topologies[index].Name // We cannot update a topology's name
-		if err := updateSingleTopology(topology); err != nil {
+	firstTopName := testData.Topologies[0].Name
+	for _, top := range testData.Topologies {
+		top.Name = firstTopName
+		if err := updateSingleTopology(top); err != nil {
 			t.Fatalf(err.Error())
 		}
 	}
@@ -176,6 +176,17 @@ func UpdateTestTopologies(t *testing.T) {
 		if err := updateSingleTopology(topology); err != nil {
 			t.Fatalf(err.Error())
 		}
+	}
+
+	// attempt to add cachegroup that doesn't meet DS required capabilities
+	top, _, err := TOSession.GetTopologyWithHdr("top-for-ds-req", nil)
+	if err != nil {
+		t.Fatalf("cannot GET topology: %v", err)
+	}
+	top.Nodes = append(top.Nodes, tc.TopologyNode{Cachegroup: "cachegroup1", Parents: []int{0}})
+	_, _, err = TOSession.UpdateTopology(top.Name, *top)
+	if err == nil {
+		t.Errorf("making invalid update to topology - expected: error, actual: nil")
 	}
 }